Memory tests

+++ src/main/java/tests/WhileShort.java
@@ -2,7 +2,7 @@ package tests;
 public class WhileShort extends perftests.AbstractTest {
      public void run() {
          short i = 0;
-         int d = Short.MAX_VALUE;
+         int d = Integer.MAX_VALUE;
          while (true) {
              i++;
              if (i == Short.MAX_VALUE) {
